Revert "Reintroduce "Improve HybridApp initialProps""
