+++ src/components/InitialURLContextProvider.tsx
@@ -1,8 +1,15 @@
+import {findFocusedRoute} from '@react-navigation/native';
 import React, {createContext, useEffect, useMemo, useState} from 'react';
 import type {ReactNode} from 'react';
 import {Linking} from 'react-native';
+import {signInAfterTransitionFromOldDot} from '@libs/actions/Session';
+import Navigation, {navigationRef} from '@navigation/Navigation';
 import type {AppProps} from '@src/App';
+import CONST from '@src/CONST';
 import type {Route} from '@src/ROUTES';
+import ROUTES from '@src/ROUTES';
+import SCREENS from '@src/SCREENS';
+import {useSplashScreenStateContext} from '@src/SplashScreenStateContext';
 
 type InitialUrlContextType = {
     initialURL: Route | undefined;
@@ -20,18 +27,40 @@ type InitialURLContextProviderProps = AppProps & {
     children: ReactNode;
 };
 
-function InitialURLContextProvider({children, url}: InitialURLContextProviderProps) {
+function InitialURLContextProvider({children, url, hybridAppSettings, timestamp}: InitialURLContextProviderProps) {
     const [initialURL, setInitialURL] = useState<Route | undefined>();
+    const {splashScreenState, setSplashScreenState} = useSplashScreenStateContext();
 
     useEffect(() => {
-        if (url) {
-            setInitialURL(url);
+        if (url && hybridAppSettings) {
+            signInAfterTransitionFromOldDot(hybridAppSettings).then(() => {
+                setInitialURL(url);
+
+                const parsedUrl = Navigation.parseHybridAppUrl(url);
+
+                Navigation.isNavigationReady().then(() => {
+                    if (parsedUrl.startsWith(`/${ROUTES.SHARE_ROOT}`)) {
+                        const focusRoute = findFocusedRoute(navigationRef.getRootState());
+                        if (focusRoute?.name === SCREENS.SHARE.SHARE_DETAILS || focusRoute?.name === SCREENS.SHARE.SUBMIT_DETAILS) {
+                            Navigation.goBack(ROUTES.SHARE_ROOT);
+                            return;
+                        }
+                    }
+                    Navigation.navigate(parsedUrl);
+                });
+
+                if (splashScreenState === CONST.BOOT_SPLASH_STATE.HIDDEN) {
+                    return;
+                }
+                setSplashScreenState(CONST.BOOT_SPLASH_STATE.READY_TO_BE_HIDDEN);
+            });
             return;
         }
         Linking.getInitialURL().then((initURL) => {
             setInitialURL(initURL as Route);
         });
-    }, [url]);
+        // eslint-disable-next-line react-compiler/react-compiler, react-hooks/exhaustive-deps
+    }, [url, hybridAppSettings, timestamp]);
 
     const initialUrlContext = useMemo(
         () => ({
