--- src/setup/hybridApp/index.ts
@@ -1,2 +1,0 @@
-// This file is intentionally empty as Hybrid App is a native mobile platforms feature
-// See `index.native.ts` for the native implementation.
